updated location for signup and helpers
